requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
pyTelegramBotAPI==4.17.0
python-dotenv==1.0.1
requests==2.31.0
telebot==0.0.5
urllib3==2.2.1
blinker==1.7.0
click==8.1.7
Flask==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
Werkzeug==3.0.2
APScheduler==3.6.3
cachetools==4.2.2
python-telegram-bot==13.7
pytz==2024.1
six==1.16.0
tornado==6.4
tzlocal==5.2
